getAngleFromSource() returns int16t and not float + changed radius define

+++ ProjetCodeSource/CodeSource/audio_processing.h
@@ -27,6 +27,6 @@ void wait_send_to_computer(void);
 float* get_audio_buffer_ptr(BUFFER_NAME_t name);
 
 //Returns the angle (in deg) from the center of the robot and a noise source
-float getAngleFromSource(void);
+int16_t getAngleFromSource(void);
 
 #endif /* AUDIO_PROCESSING_H */
